Provided Global Fees, retrieve and update functionality.

+++ backend/src/main/java/com/bvcott/booking/config/GlobalFeeConfig.java
@@ -1,7 +1,5 @@
 package com.bvcott.booking.config;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.Min;
@@ -12,9 +10,9 @@ public class GlobalFeeConfig {
     @Id private int id = 1;
 
     @Min(0)
-    private BigDecimal baseCharge;
+    private double baseCharge;
     @Min(0)
-    private BigDecimal chargePerRoom;
+    private double chargePerRoom;
     @Min(0)
-    private BigDecimal transactionFee;
+    private double transactionFee;
 }
